pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kimiconfig"
version = "0.2.4"
description = "Гибкая система управления конфигурацией для Python-приложений"
readme = "README.md"
authors = [
    { name = "KimiFish", email = "author@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "kimiutils>=0.1.0",
]

[project.optional-dependencies]
full = [
    "rich>=10.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[project.urls]
Repository = "https://github.com/username/kimiconfig"
Documentation = "https://github.com/username/kimiconfig#readme"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88 